Added training pipeline
Look4LTRs relies on a machine learning component and downstream parameters are tied to the results of the model. If the user has genomes with LTR-retrotransposons not well depicted in Repbase/literature, then they may want to retrain the model on these values. These files aid the user in automatically training a model and tuning down-stream parameters given the chromosomes of their genomes as well as LTR-retrotransposon locations.